Fix infinite loop on simplification producing already existing segment.
Special case happening with poor quality shapefile when the line shared
between 2 shapes are not exactly the same (not all points shared), the
result of simplifying a segment can be an already existing segment.

Deal also with the case of a very narrow loop, simplified to the same
and unique segment on both sides, if not connected to anything else
this single segment will be removed.